Added processing of TextPad registry entries
In older versions of TextPad, the registry is used to store some of its settings.  This version exports and imports those registry settings as well as any configuration files used in newer versions.  You must run the application as an administrator for older versions of TextPad to allow it to process the registry settings.